I have made significant changes to the import procedures. All the procedures are now automated and do not rely on loops. Instead, they use "Map" to assign expressions imported from libraries with "Get" to expressions. The new changes allow users to select which libraries they want to import.
When using simple models, the user can choose the order in which libraries are imported. However, due to the many variables involved, the same functionality is unavailable for Horndeski interaction. Each library is marked by the number of fields, the number of kinetic terms, and the order of gravity perturbation theory. Further, I want to allow users to choose the number of scalar fields participating in the interaction and import libraries up to a given order.